+++ src/sub_array.cpp
@@ -8,14 +8,13 @@ class MinimalSubscriber : public rclcpp::Node
 {
 public:
   MinimalSubscriber()
-  : Node("array_sub"),
-  best_effort(rclcpp::KeepLast(10))
+      : Node("array_sub"),
+        best_effort(rclcpp::KeepLast(10))
 
   {
     subscription_ = this->create_subscription<ros2labview_examples::msg::Intarray>(
-      "intarr", best_effort.reliability(be), [this](const ros2labview_examples::msg::Intarray::SharedPtr msg){
-      RCLCPP_INFO(this->get_logger(), "I heard: %d %d %d", msg->num[0],msg->num[1],msg->num[2]);
-    });
+        "intarr", best_effort.reliability(be), [this](const ros2labview_examples::msg::Intarray::SharedPtr msg)
+        { RCLCPP_INFO(this->get_logger(), "I heard: %d %d %d", (int)msg->num[0], (int)msg->num[1], (int)msg->num[2]); });
   }
 
 private:
@@ -24,7 +23,7 @@ private:
   rclcpp::QoS best_effort;
 };
 
-int main(int argc, char * argv[])
+int main(int argc, char *argv[])
 {
   rclcpp::init(argc, argv);
   rclcpp::spin(std::make_shared<MinimalSubscriber>());
