+++ src/sub_cluster.cpp
@@ -8,13 +8,12 @@ class MinimalSubscriber : public rclcpp::Node
 {
 public:
   MinimalSubscriber()
-  : Node("cluster_sub"),
-    best_effort(rclcpp::KeepLast(10))
+      : Node("cluster_sub"),
+        best_effort(rclcpp::KeepLast(10))
   {
     subscription_ = this->create_subscription<ros2labview_examples::msg::Clusterdemo>(
-      "intcust", best_effort.reliability(be), [this](const ros2labview_examples::msg::Clusterdemo::SharedPtr msg){
-      RCLCPP_INFO(this->get_logger(), "I heard: %d and %d", msg->x,msg->y);
-    });
+        "intcust", best_effort.reliability(be), [this](const ros2labview_examples::msg::Clusterdemo::SharedPtr msg)
+        { RCLCPP_INFO(this->get_logger(), "I heard: %d and %d", (int)msg->x, (int)msg->y); });
   }
 
 private:
@@ -23,7 +22,7 @@ private:
   rclcpp::QoS best_effort;
 };
 
-int main(int argc, char * argv[])
+int main(int argc, char *argv[])
 {
   rclcpp::init(argc, argv);
   rclcpp::spin(std::make_shared<MinimalSubscriber>());
