+++ src/sub_custom_int.cpp
@@ -8,14 +8,13 @@ class MinimalSubscriber : public rclcpp::Node
 {
 public:
   MinimalSubscriber()
-  : Node("cint_sub"),
-  best_effort(rclcpp::KeepLast(10))
+      : Node("cint_sub"),
+        best_effort(rclcpp::KeepLast(10))
 
   {
     subscription_ = this->create_subscription<ros2labview_examples::msg::Myint>(
-      "clabint", best_effort.reliability(be), [this](const ros2labview_examples::msg::Myint::SharedPtr msg){
-      RCLCPP_INFO(this->get_logger(), "I heard: %d", msg->num);
-    });
+        "clabint", best_effort.reliability(be), [this](const ros2labview_examples::msg::Myint::SharedPtr msg)
+        { RCLCPP_INFO(this->get_logger(), "I heard: %d", (int)msg->num); });
   }
 
 private:
@@ -24,7 +23,7 @@ private:
   rclcpp::QoS best_effort;
 };
 
-int main(int argc, char * argv[])
+int main(int argc, char *argv[])
 {
   rclcpp::init(argc, argv);
   rclcpp::spin(std::make_shared<MinimalSubscriber>());
